Refactor Model, SectorModel
- move inputs and parameters to SectorModel (ScenarioModel doesn't use this)
- remove CompositeModel (not used, other that for SosModel, which is not possible
  to 'simulate' or define inputs for
- remove add_dependency (to be moved up to SosModel)
