+++ src/smif/model/sector_model.py
@@ -34,7 +34,6 @@ The key functions include:
   approaches
 
 """
-import logging
 import sys
 from abc import ABCMeta, abstractmethod
 
@@ -88,17 +87,53 @@ class SectorModel(Model, metaclass=ABCMeta):
     """
     def __init__(self, name):
         super().__init__(name)
+        self._inputs = {}
+        self._parameters = {}
 
-        self.initial_conditions = []
-        self._interventions = {}
+    @property
+    def inputs(self):
+        """All model inputs defined at this layer
 
-        self.logger = logging.getLogger(__name__)
+        Returns
+        -------
+        dict of {input_name: smif.metadata.Spec}
+        """
+        return self._inputs
+
+    @property
+    def parameters(self):
+        """Model parameters
+
+        Returns
+        -------
+        dict of {parameter_name: smif.metadata.Spec}
+        """
+        return self._parameters
+
+    def add_input(self, spec):
+        """Add an input
+
+        Arguments
+        ---------
+        spec: smif.metadata.Spec
+        """
+        self.inputs[spec.name] = spec
+
+    def add_parameter(self, spec):
+        """Add a parameter
+
+        Arguments
+        ---------
+        spec: smif.metadata.Spec
+        """
+        self.parameters[spec.name] = spec
 
     @classmethod
     def from_dict(cls, config):
+        """Create object from dictionary serialisation
+        """
         model = cls(config['name'])
         model.description = config['description']
-        model.path = config['path']
         for input_ in config['inputs']:
             model.add_input(Spec.from_dict(input_))
         for output in config['outputs']:
